+++ index/include/indri/TermList.hpp
@@ -111,20 +111,20 @@ namespace indri {
         indri::utility::RVLCompressStream out( buffer );
         
         // write count of terms and fields in the document first
-        int termCount = _terms.size();
-        int fieldCount = _fields.size();
+        int termCount = (int)_terms.size();
+        int fieldCount = (int)_fields.size();
 
         out << termCount
             << fieldCount;
         
         // write out terms
-        for( int i=0; i<_terms.size(); i++ ) {
+        for( size_t i=0; i<_terms.size(); i++ ) {
           assert( _terms[i] >= 0 );
           out << _terms[i];
         }
 
         // write out fields
-        for( unsigned int i=0; i<_fields.size(); i++ ) {
+        for( size_t i=0; i<_fields.size(); i++ ) {
 
           assert( _fields[i].id >= 0 );
           assert( _fields[i].ordinal == i + 1 );
